Comunicação entre microsservicos fabrica e loja.

+++ fabrica-nike/src/main/java/br/com/josivanesoftware/notaFiscal/model/NotaFiscal.java
@@ -9,6 +9,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -25,5 +26,5 @@ public class NotaFiscal {
     private LocalDateTime dataEmissao;
     private String origem;
     private String destino;
-    private Double valorTotal;
+    private BigDecimal valorTotal;
 }
